requirements.txt
langchain-core~=0.3.15
langchain-community~=0.3.5
langchain~=0.3.7
langchain-openai~=0.2.6
pydantic
openai
elasticsearch
python-dotenv~=1.0.1
psycopg2-binary
sentence-transformers
tiktoken
setuptools~=68.2.0
